requirements.txt
akshare>=1.12.0
pyyaml>=6.0
schedule>=1.2.0
pandas>=2.0.0
